requirements.txt
langchain
langchain-community
langchain-openai 
python-dotenv
reportlab  
requests  
httpx